🗺️ Atlas: Extract AppState to state.rs module
