+++ autumn/src/lib.rs
@@ -533,248 +533,12 @@ pub mod reexports {
     pub use validator;
 }
 
-/// Shared application state passed to all route handlers.
-///
-/// Holds framework-managed resources such as the database connection pool.
-/// Axum requires handler state to be [`Clone`], so internal resources use
-/// `Arc` or are already cheaply cloneable (`deadpool::Pool` is `Arc`-wrapped
-/// internally).
-///
-/// This struct is normally constructed by [`app::AppBuilder::run`] and
-/// should not need to be created manually. It is public so that custom
-/// Axum extractors can access framework resources via
-/// `State<AppState>`.
-///
-/// # Examples
-///
-/// ```rust
-/// use autumn_web::AppState;
-///
-/// // State without a database (e.g., for testing)
-/// let state = AppState {
-///     pool: None,
-///     profile: Some("dev".into()),
-///     started_at: std::time::Instant::now(),
-///     health_detailed: true,
-///     metrics: autumn_web::middleware::MetricsCollector::new(),
-///     log_levels: autumn_web::actuator::LogLevels::new("info"),
-///     task_registry: autumn_web::actuator::TaskRegistry::new(),
-///     config_props: Default::default(),
-/// };
-/// ```
-#[derive(Clone)]
-pub struct AppState {
-    /// Database connection pool, or `None` when no `database.url` is configured.
-    #[cfg(feature = "db")]
-    pub pool:
-        Option<diesel_async::pooled_connection::deadpool::Pool<diesel_async::AsyncPgConnection>>,
-
-    /// Active profile name (e.g., "dev", "prod", "staging").
-    pub profile: Option<String>,
-
-    /// When the application started. Used for uptime calculation.
-    pub started_at: std::time::Instant,
-
-    /// Whether the health endpoint should include detailed info.
-    pub health_detailed: bool,
-
-    /// In-memory metrics collector for the `/actuator/metrics` endpoint.
-    pub metrics: middleware::MetricsCollector,
-
-    /// Runtime log level state for the `/actuator/loggers` endpoint.
-    pub log_levels: actuator::LogLevels,
-
-    /// Scheduled task registry for the `/actuator/tasks` endpoint.
-    pub task_registry: actuator::TaskRegistry,
-
-    /// Resolved config properties with source tracking for `/actuator/configprops`.
-    pub config_props: actuator::ConfigProperties,
-}
-
-impl AppState {
-    /// Returns the active profile name, or `"default"` if none is set.
-    #[must_use]
-    pub fn profile(&self) -> &str {
-        self.profile.as_deref().unwrap_or("default")
-    }
-
-    /// Returns how long the application has been running.
-    #[must_use]
-    pub fn uptime(&self) -> std::time::Duration {
-        self.started_at.elapsed()
-    }
-
-    /// Format uptime as a human-readable string (e.g., "2h 15m").
-    #[must_use]
-    pub fn uptime_display(&self) -> String {
-        let secs = self.started_at.elapsed().as_secs();
-        if secs < 60 {
-            format!("{secs}s")
-        } else if secs < 3600 {
-            format!("{}m {}s", secs / 60, secs % 60)
-        } else {
-            let hours = secs / 3600;
-            let mins = (secs % 3600) / 60;
-            format!("{hours}h {mins}m")
-        }
-    }
-
-    /// Create an `AppState` suitable for testing, with sensible defaults
-    /// for all fields. Database pool is `None`.
-    #[cfg(test)]
-    #[allow(dead_code)]
-    pub(crate) fn for_test() -> Self {
-        Self {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: None,
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        }
-    }
-}
-
-impl std::fmt::Debug for AppState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut s = f.debug_struct("AppState");
-        #[cfg(feature = "db")]
-        s.field(
-            "pool",
-            &self
-                .pool
-                .as_ref()
-                .map(|p| format!("Pool(max={})", p.status().max_size)),
-        );
-        s.field("profile", &self.profile)
-            .field("started_at", &self.started_at)
-            .field("health_detailed", &self.health_detailed)
-            .field("metrics", &"MetricsCollector")
-            .field("log_levels", &"LogLevels")
-            .field("task_registry", &"TaskRegistry")
-            .finish_non_exhaustive()
-    }
-}
+pub mod state;
+pub use state::AppState;
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn app_state_debug_without_pool() {
-        let state = AppState {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: Some("dev".into()),
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        let debug = format!("{state:?}");
-        assert!(debug.contains("AppState"));
-        assert!(debug.contains("dev"));
-    }
-
-    #[cfg(feature = "db")]
-    #[test]
-    fn app_state_debug_with_pool() {
-        let config = config::DatabaseConfig {
-            url: Some("postgres://localhost/test".into()),
-            pool_size: 5,
-            ..Default::default()
-        };
-        let pool = db::create_pool(&config).unwrap().unwrap();
-        let state = AppState {
-            pool: Some(pool),
-            profile: None,
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        let debug = format!("{state:?}");
-        assert!(debug.contains("Pool(max=5)"));
-    }
-
-    fn require_clone<T: Clone>(t: &T) -> T {
-        t.clone()
-    }
-
-    #[test]
-    fn app_state_is_clone() {
-        let state = AppState {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: None,
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        let _cloned = require_clone(&state);
-    }
-
-    #[test]
-    fn app_state_profile_accessor() {
-        let state = AppState {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: Some("staging".into()),
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        assert_eq!(state.profile(), "staging");
-    }
-
-    #[test]
-    fn app_state_profile_default() {
-        let state = AppState {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: None,
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        assert_eq!(state.profile(), "default");
-    }
-
-    #[test]
-    fn app_state_uptime_display() {
-        let state = AppState {
-            #[cfg(feature = "db")]
-            pool: None,
-            profile: None,
-            started_at: std::time::Instant::now(),
-            health_detailed: true,
-            metrics: middleware::MetricsCollector::new(),
-            log_levels: actuator::LogLevels::new("info"),
-            task_registry: actuator::TaskRegistry::new(),
-            config_props: actuator::ConfigProperties::default(),
-        };
-        let display = state.uptime_display();
-        assert!(
-            display.contains('s'),
-            "uptime should contain 's': {display}"
-        );
-    }
 
     #[test]
     fn app_fn_creates_builder() {
